Move artifacts back so that they can be packaged properly
