+++ packages/pcd/rln-pcd/src/RLNPCD.ts
@@ -20,7 +20,7 @@ import { Identity } from "@semaphore-protocol/identity";
 import JSONBig from "json-bigint";
 import { Proof, RLN, RLNFullProof } from "rlnjs";
 import { v4 as uuid } from "uuid";
-import verificationKeyJSON from "./artifacts/16.json";
+import verificationKeyJSON from "./16.json";
 
 let initArgs: RLNPCDInitArgs | undefined = undefined;
