+++ packages/pcd/zk-eddsa-event-ticket-pcd/artifacts/circuit.zkey
