+++ packages/pcd/zk-eddsa-frog-pcd/artifacts/circuit.zkey
