+++ packages/pcd/zk-eddsa-frog-pcd/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "@pcd/tsconfig/ts-library.json",
-  "include": ["src", "src/artifacts/*.json"],
+  "include": ["src", "src/*.json"],
   "exclude": ["dist", "build", "node_modules"],
   "references": [
     {
